pyproject.toml
[tool.poetry]
name = "sixgpt-miner"
version = "0.1.0"
description = "GPT Miner"
authors = ["Flotar Haven <flotarhaven@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
httpx = "^0.27.0"
click = "^8.1.7"
google-api-python-client = "^2.138.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
